Move CompressARM64CFI to NativeAOT-specific CorInfoImpl
Unwind info is generated as CFI for NativeAOT only.
